Update nslookup_ping_host_whois
This script was for the full intention of Network Admin,Engineers for troubleshooting purposes. 

Plug N Play = Copy and paste all of the hosts "Line per line". Once done hit enter to add an extra empty line and hit CTRL D to quit the interactive prompt and then perform the action you would like to take.